+++ lib/ReactViews/Map/Navigation/augmented_virtuality_tool.scss.d.ts
@@ -12,13 +12,10 @@ interface CssExports {
   'btnPrimaryHover': string;
   'charcoalGrey': string;
   'colorPrimary': string;
-  'colorSplitter': string;
   'dark': string;
   'darkWithOverlay': string;
   'mapButtonColor': string;
   'modalHighlight': string;
-  'textDarker': string;
-  'textLight': string;
   'toolButton': string;
   'with-time-series-controls': string;
   'withTimeSeriesControls': string;