+++ lib/ReactViews/Search/search-result.scss.d.ts
@@ -5,11 +5,14 @@ interface CssExports {
   'arrowIcon': string;
   'btn': string;
   'btnLocationName': string;
+  'btnWithBorderBottom': string;
   'charcoalGrey': string;
   'colorPrimary': string;
   'colorSplitter': string;
   'dark': string;
   'darkWithOverlay': string;
+  'grey': string;
+  'greyLightest': string;
   'icon': string;
   'light': string;
   'mapButtonColor': string;