+++ lib/ReactViews/Workbench/Controls/colorscalerange-section.scss.d.ts
@@ -2,17 +2,8 @@
 // Please do not change this file!
 interface CssExports {
   'btn': string;
-  'charcoalGrey': string;
-  'colorPrimary': string;
-  'colorSplitter': string;
   'colorscalerange': string;
-  'dark': string;
-  'darkWithOverlay': string;
   'field': string;
-  'mapButtonColor': string;
-  'modalHighlight': string;
-  'textDarker': string;
-  'textLight': string;
   'title': string;
 }
 declare var cssExports: CssExports;